Temporarly finished worked with implementing 32-bit addressing. More refactor - making files cleaner, easier to maintain, fixing code (if needed).

+++ os/kernel/src/v8086/mod_rm_parsing.h
@@ -9,7 +9,7 @@ uint32_t* get_dword_register(v8086* machine, uint8_t reg_field);
 void* get_variable_length_register(v8086* machine, uint8_t reg_field, uint8_t width);
 
 uint16_t* select_segment_register(v8086* machine, segment_register_select select);
-int16_t calculate_segment_offset_from_mode(v8086* machine, uint8_t mod_rm, uint16_t* segment, uint16_t* offset);
+int16_t calculate_segment_offset_from_mode(v8086* machine, uint8_t mod_rm, uint16_t* segment, uint32_t* offset);
 int16_t calculate_segment_offset_from_mode_32(v8086* machine, uint8_t mod_rm, uint16_t* segment, uint32_t* offset);
 void* get_memory_from_mode(v8086* machine, uint8_t mod_rm, uint8_t width);
